% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_7.R
\name{f}
\alias{f}
\title{A helper function for implementation of team_7}
\usage{
f(dframe)
}
\arguments{
\item{dframe}{A data frame}
}
\value{
A data frame
}
\description{
A helper function for implementation of team_7
}
