% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_7.R
\name{plot_team_7}
\alias{plot_team_7}
\title{Plot code for object returned from team_7}
\usage{
plot_team_7(ozplus)
}
\arguments{
\item{ozplus}{an object returned from team_7()}
}
\description{
Plot code for object returned from team_7
}
\examples{
plot_team_7(team_7())
}
