% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_team8.R
\name{plot_team_8}
\alias{plot_team_8}
\title{Returns a ggplot from the output of the team_8 function}
\usage{
plot_team_8(shape)
}
\arguments{
\item{shape}{shape data frame output from the team_8 function}
}
\description{
Returns a ggplot from the output of the team_8 function
}
