% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_5.R
\name{team_5}
\alias{team_5}
\title{Welcome to the function, team_5}
\usage{
team_5(file, tolerance)
}
\arguments{
\item{file}{a character string of your shape file}

\item{tolerance}{the value used for thinning the polygona; a single number of the tolerance value in the metric of the input object for an innerfunction called thinnedSpatialPoly{maptools}}
}
\value{
The return value is a data frame
}
\description{
team_5 returns a data frame of the geographic information of the polygons and the additional information such as name of the country, name of the territory/state.
}
\examples{
ozfilepath <- system.file("extdata/gadm36_AUS_1.shp", ".shp", package = "mypackage")
res <- team_5(ozfilepath,0.1)
head(res)
tail(res)
unique(res$territory.state)
}
