% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_7.R
\name{team_7}
\alias{team_7}
\title{Implementation of solution from team 7}
\usage{
team_7(file = "./data/gadm36_AUS_shp/gadm36_AUS_1.shp",
  tolerance = 0.1)
}
\arguments{
\item{file}{A country's shapefile}

\item{tolerance}{A tolerance value}
}
\value{
A data frame of geographic information
}
\description{
Implementation of solution from team 7
}
