% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_8.R
\name{team_8}
\alias{team_8}
\title{Creates a data frame with group-indicated lat/long data}
\usage{
team_8(file, tolerance = 0.1)
}
\arguments{
\item{file}{a file path string}

\item{tolerance}{the tolerance value in the metric of the input object}
}
\description{
From an input .shp file, will create a data frame with columns indicating the country and state/territory of which the data represent, latitude and longitude data, and the group to which the lat/long data belong. Perfectly set-up to put into a ggplot call.
}
\examples{
#oz <- team_8("tests/testData/team_8_testData.shp",tolerance=.1)
#oz \%>\%
# ggplot(aes(x=long,y=lat,group=group)) + geom_path()

}
