% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_1.R
\name{team_1}
\alias{team_1}
\title{Wrapping the solution from team 1 into a function}
\usage{
team_1(file, tolerance)
}
\arguments{
\item{file}{A file path to a shape file}

\item{tolerance}{The tolerance value used for thinning the polygon}
}
\value{
A data frame of geographic information
}
\description{
Wrapping the solution from team 1 into a function
}
\examples{
ozfile <- system.file("extdata/ashmore_cartierIslands.shp",".shp",package=lab3Group9)
oz <- team_1(ozfile,tolerance=.1)
}
